\name{amphib}
\alias{amphib}
\docType{data}
\title{
Amphibian presence/absence data
}
\description{
}
\usage{data("amphib")}
\format{
  A data frame with 273 rows (localities) and 18 columns (amphibian species), with values representing their presence (1) or absence (0).
}
\details{
}
\source{
}
\references{
}
\examples{
data(amphib)
head(amphib)
}
\keyword{datasets}
