\name{simil}
\alias{simil}
\docType{data}
\title{
Similarity matrix among amphibian presence/absence patterns
}
\description{
}
\usage{data("simil")}
\format{
  A data frame with the similarity values among the presence/absence patterns of the 18 species of amphibians included in \code{\link{amphib}}.
}
\details{
}
\source{
}
\references{
}
\examples{
data(simil)
}
\keyword{datasets}
